+++ skillfarm/api/skillfarm.py
@@ -16,9 +16,6 @@ from django.utils.translation import gettext_lazy as _
 from allianceauth.authentication.models import UserProfile
 from allianceauth.services.hooks import get_extension_logger
 
-# Alliance Auth (External Libs)
-from app_utils.logging import LoggerAddTag
-
 # AA Skillfarm
 from skillfarm import __title__
 from skillfarm.api.helpers.core import (
@@ -30,8 +27,8 @@ from skillfarm.api.helpers.core import (
     generate_status_icon_html,
     generate_toggle_notification_button,
     get_alts_queryset,
-    get_character,
-    get_main_character,
+    get_auth_character_or_main,
+    get_skillfarm_character,
 )
 from skillfarm.api.helpers.skilldetails import (
     _calculate_sum_progress_bar,
@@ -45,8 +42,9 @@ from skillfarm.models.skillfarmaudit import (
     SkillFarmAudit,
     SkillFarmSetup,
 )
+from skillfarm.providers import AppLogger
 
-logger = LoggerAddTag(get_extension_logger(__name__), __title__)
+logger = AppLogger(my_logger=get_extension_logger(__name__), prefix=__title__)
 
 
 class OverviewSchema(Schema):
@@ -189,7 +187,7 @@ class SkillFarmApiEndpoints:
                             template_name=template,
                             context={
                                 "url": reverse(
-                                    "skillfarm:skillfarm",
+                                    "skillfarm:index",
                                     kwargs={
                                         "character_id": character.main_character.character_id
                                     },
@@ -237,7 +235,7 @@ class SkillFarmApiEndpoints:
                 f"User {request.user} requested SkillFarm details for character ID {character_id}."
             )
             # Get Main Character and check permissions
-            perm, main = get_main_character(request, character_id)
+            perm, main = get_auth_character_or_main(request, character_id)
 
             # Check permissions
             if perm is False:
@@ -343,7 +341,7 @@ class SkillFarmApiEndpoints:
                 f"User {request.user} requested SkillFarm skill setup for character ID {character_id}."
             )
             # Get Main Character and check permissions
-            perm, character = get_character(request, character_id)
+            perm, character = get_skillfarm_character(request, character_id)
 
             if perm is False:
                 logger.warning(
@@ -377,7 +375,7 @@ class SkillFarmApiEndpoints:
                 f"User {request.user} requested SkillFarm skill info for character ID {character_id}."
             )
             # Get Main Character and check permissions
-            perm, character = get_character(request, character_id)
+            perm, character = get_skillfarm_character(request, character_id)
 
             if perm is False:
                 logger.warning(