+++ skillfarm/tests/test_models.py
@@ -4,63 +4,77 @@ from django.utils import timezone
 
 # AA Skillfarm
 from skillfarm.models.skillfarmaudit import (
-    CharacterSkill,
-    CharacterSkillqueueEntry,
     SkillFarmAudit,
-    SkillFarmSetup,
 )
-from skillfarm.tests.testdata.allianceauth import load_allianceauth
-from skillfarm.tests.testdata.eveuniverse import load_eveuniverse
-from skillfarm.tests.testdata.skillfarm import (
-    create_skillfarm_character,
+from skillfarm.tests import SkillFarmTestCase
+from skillfarm.tests.testdata.integrations.allianceauth import load_allianceauth
+from skillfarm.tests.testdata.utils import (
+    create_skillfarm_character_from_user,
     create_update_status,
 )
 
 MODULE_PATH = "skillfarm.models.skillfarmaudit"
 
 
-class TestSkillfarmModel(TestCase):
+class TestSkillfarmModel(SkillFarmTestCase):
     @classmethod
     def setUpClass(cls):
         super().setUpClass()
-        load_allianceauth()
 
-        cls.audit = create_skillfarm_character(1001)
+        # Create SkillfarmAudit instance
+        cls.skillfarm_audit = create_skillfarm_character_from_user(cls.user)
 
     def test_should_return_string_audit(self):
-        """Test should return the Audit Character Data"""
-        self.assertEqual(str(self.audit), "Gneuten - Active: True - Status: incomplete")
+        """
+        Test should return string representation of SkillFarmAudit.
+        """
+        self.assertEqual(
+            str(self.skillfarm_audit), "Gneuten - Active: True - Status: incomplete"
+        )
 
     def test_should_return_esi_scopes(self):
-        """Test should return the ESI Scopes for Skillfarm"""
+        """
+        Test should return ESI scopes required for SkillFarmAudit.
+        """
         self.assertEqual(
-            self.audit.get_esi_scopes(),
+            self.skillfarm_audit.get_esi_scopes(),
             ["esi-skills.read_skills.v1", "esi-skills.read_skillqueue.v1"],
         )
 
     def test_is_cooldown_should_return_false(self):
-        """Test should return False for is_cooldown Property"""
-        self.assertFalse(self.audit.is_cooldown)
+        """
+        Test should return False for is_cooldown Property.
+        """
+        self.assertFalse(self.skillfarm_audit.is_cooldown)
 
     def test_is_cooldown_should_return_true(self):
-        """Test should return True for is_cooldown Property"""
-        self.audit.last_notification = timezone.now()
-        self.assertTrue(self.audit.is_cooldown)
+        """
+        Test should return True for is_cooldown Property.
+        """
+        self.skillfarm_audit.last_notification = timezone.now()
+        self.assertTrue(self.skillfarm_audit.is_cooldown)
 
     def test_last_update_should_return_incomplete(self):
-        """Test should return incomplete description for last_update Property"""
+        """
+        Test should return incomplete for last_update Property.
+        """
         self.assertEqual(
-            self.audit.last_update, "One or more sections have not been updated"
+            self.skillfarm_audit.last_update,
+            "One or more sections have not been updated",
         )
 
     def test_reset_has_token_error_should_return_false(self):
-        """Test should reset has_token_error"""
-        self.assertFalse(self.audit.reset_has_token_error())
+        """
+        Test should not reset has_token_error.
+        """
+        self.assertFalse(self.skillfarm_audit.reset_has_token_error())
 
     def test_reset_has_token_error_should_return_true(self):
-        """Test should reset has_token_error"""
+        """
+        Test should reset has_token_error.
+        """
         create_update_status(
-            self.audit,
+            self.skillfarm_audit,
             section=SkillFarmAudit.UpdateSection.SKILLQUEUE,
             is_success=False,
             error_message="",
@@ -70,4 +84,4 @@ class TestSkillfarmModel(TestCase):
             last_update_at=timezone.now(),
             last_update_finished_at=timezone.now(),
         )
-        self.assertTrue(self.audit.reset_has_token_error())
+        self.assertTrue(self.skillfarm_audit.reset_has_token_error())