+++ skillfarm/tests/test_tasks.py
@@ -2,8 +2,9 @@
 from unittest.mock import patch
 
 # Django
+from django.db import models
 from django.db.utils import Error
-from django.test import TestCase, override_settings
+from django.test import override_settings
 from django.utils import timezone
 
 # Alliance Auth
@@ -11,33 +12,40 @@ from allianceauth.authentication.models import UserProfile
 
 # AA Skillfarm
 from skillfarm import tasks
+from skillfarm.models.prices import EveTypePrice
 from skillfarm.models.skillfarmaudit import SkillFarmAudit
-from skillfarm.tests.testdata.allianceauth import load_allianceauth
-from skillfarm.tests.testdata.eveuniverse import load_eveuniverse
-from skillfarm.tests.testdata.skillfarm import (
-    add_skillfarmaudit_character_to_user,
-    create_evetypeprice,
-    create_skill_character,
-    create_skillfarm_character,
+from skillfarm.tests import NoSocketsTestCase, SkillFarmTestCase
+from skillfarm.tests.testdata.utils import (
+    create_character_skill,
+    create_eve_type_price,
+    create_skillfarm_character_from_user,
     create_skillsetup_character,
     create_update_status,
-    create_user_from_evecharacter_with_access,
 )
 
 TASK_PATH = "skillfarm.tasks"
 
 
+@override_settings(
+    CELERY_ALWAYS_EAGER=True,
+    CELERY_EAGER_PROPAGATES_EXCEPTIONS=True,
+)
 @patch(TASK_PATH + ".update_character", spec=True)
-class TestUpdateAllSkillfarm(TestCase):
+class TestUpdateAllSkillfarm(SkillFarmTestCase):
+    """Test the update_all_skillfarm task."""
+
     @classmethod
     def setUpClass(cls):
         super().setUpClass()
-        load_allianceauth()
-        load_eveuniverse()
 
-        cls.audit = create_skillfarm_character(1001)
+        cls.skillfarm_audit = create_skillfarm_character_from_user(cls.user)
 
     def test_should_update_all_skillfarm(self, mock_update_all_skillfarm):
+        """
+            Test should start update_character for each SkillFarmAudit.
+        :return:
+        :rtype:
+        """
         # when
         tasks.update_all_skillfarm()
         # then
@@ -47,32 +55,37 @@ class TestUpdateAllSkillfarm(TestCase):
 @override_settings(
     CELERY_ALWAYS_EAGER=True,
     CELERY_EAGER_PROPAGATES_EXCEPTIONS=True,
-    APP_UTILS_OBJECT_CACHE_DISABLED=True,
 )
 @patch(TASK_PATH + ".chain", spec=True)
 @patch(TASK_PATH + ".logger", spec=True)
-class TestUpdateCharacter(TestCase):
+class TestUpdateCharacter(SkillFarmTestCase):
+    """Test the update_character task."""
+
     @classmethod
     def setUpClass(cls):
         super().setUpClass()
-        load_allianceauth()
-        load_eveuniverse()
 
-        cls.audit = create_skillfarm_character(1001)
+        cls.skillfarm_audit = create_skillfarm_character_from_user(cls.user)
 
     def test_update_character_should_no_updated(self, mock_logger, __):
+        """
+        Test should not update character if no updates are needed.
+        """
         # when
-        tasks.update_character(self.audit.pk)
+        tasks.update_character(self.skillfarm_audit.pk)
         # then
         mock_logger.info.assert_called_once_with(
             "No updates needed for %s",
-            self.audit.character.character_name,
+            self.skillfarm_audit.character.character_name,
         )
 
     def test_update_character_should_update(self, mock_logger, mock_chain):
+        """
+        Test should update character if updates are needed.
+        """
         # given
         create_update_status(
-            self.audit,
+            self.skillfarm_audit,
             section=SkillFarmAudit.UpdateSection.SKILLS,
             is_success=True,
             error_message="",
@@ -84,7 +97,7 @@ class TestUpdateCharacter(TestCase):
         )
 
         # when
-        tasks.update_character(self.audit.pk)
+        tasks.update_character(self.skillfarm_audit.pk)
         # then
         mock_chain.assert_called_once()
 
@@ -93,33 +106,33 @@ class TestUpdateCharacter(TestCase):
 @override_settings(
     CELERY_ALWAYS_EAGER=True,
     CELERY_EAGER_PROPAGATES_EXCEPTIONS=True,
-    APP_UTILS_OBJECT_CACHE_DISABLED=True,
 )
-class TestCheckSkillfarmNotification(TestCase):
+class TestCheckSkillfarmNotification(SkillFarmTestCase):
     @classmethod
     def setUpClass(cls):
         super().setUpClass()
-        load_allianceauth()
-        load_eveuniverse()
 
-        cls.user, cls.character_ownership = create_user_from_evecharacter_with_access(
-            1001
+        cls.skillfarm_audit = create_skillfarm_character_from_user(cls.user)
+        cls.skillfarm_audit_2 = create_skillfarm_character_from_user(
+            cls.no_permission_user
         )
-        cls.user2, cls.character_ownership2 = create_user_from_evecharacter_with_access(
-            1002
-        )
-        cls.audit = add_skillfarmaudit_character_to_user(cls.user, 1001)
-        cls.audit2 = add_skillfarmaudit_character_to_user(cls.user2, 1002)
-        cls.audit3 = add_skillfarmaudit_character_to_user(cls.user2, 1003)
+        cls.skillfarm_audit_3 = create_skillfarm_character_from_user(cls.superuser)
 
-    def _set_notifiaction_status(self, audits, status):
+    def _set_notification_status(
+        self, audits: models.QuerySet[SkillFarmAudit], status: bool
+    ):
+        """Set notification status for SkillFarmAudit."""
         for audit in audits:
             audit.notification = status
             audit.save()
 
     def test_no_notification_should_return_false(self, mock_audit_filter):
-        audits = [self.audit, self.audit2]
-        self._set_notifiaction_status(audits, False)
+        """
+        Test should not send notification if notification is disabled.
+        """
+        # given
+        audits = [self.skillfarm_audit, self.skillfarm_audit_2]
+        self._set_notification_status(audits, False)
         mock_audit_filter.return_value = audits
         # when
         tasks.check_skillfarm_notifications()
@@ -131,8 +144,11 @@ class TestCheckSkillfarmNotification(TestCase):
     def test_notifiaction_with_no_skillsetup_should_return_false(
         self, mock_audit_filter
     ):
-        audits = [self.audit, self.audit2, self.audit3]
-        self._set_notifiaction_status(audits, True)
+        """
+        Test should not send notification if no SkillFarmSetup is found.
+        """
+        audits = [self.skillfarm_audit, self.skillfarm_audit_2, self.skillfarm_audit_3]
+        self._set_notification_status(audits, True)
         mock_audit_filter.return_value = audits
         # when
         tasks.check_skillfarm_notifications()
@@ -141,59 +157,23 @@ class TestCheckSkillfarmNotification(TestCase):
             self.assertFalse(audit.notification_sent)
             self.assertIsNone(audit.last_notification)
 
-
-@patch(TASK_PATH + ".SkillFarmAudit.objects.filter", spec=True)
-@override_settings(
-    CELERY_ALWAYS_EAGER=True,
-    CELERY_EAGER_PROPAGATES_EXCEPTIONS=True,
-    APP_UTILS_OBJECT_CACHE_DISABLED=True,
-)
-class TestCheckSkillfarmNotificationSuccess(TestCase):
-    @classmethod
-    def setUpClass(cls):
-        super().setUpClass()
-        load_allianceauth()
-        load_eveuniverse()
-
-        cls.user, cls.character_ownership = create_user_from_evecharacter_with_access(
-            1001
+    def test_notification_should_return_true(self, mock_audit_filter):
+        """
+        Test should send notification if notification is enabled and SkillFarmSetup exists.
+        """
+        # given
+        create_skillsetup_character(
+            character_id=self.skillfarm_audit.character.character_id,
+            skillset=["skill1"],
         )
-        cls.audit = add_skillfarmaudit_character_to_user(cls.user, 1001)
-        cls.user2, cls.character_ownership2 = create_user_from_evecharacter_with_access(
-            1002
+        create_character_skill(
+            character_id=self.skillfarm_audit.character.character_id, evetype_id=1
         )
-        cls.audit2 = add_skillfarmaudit_character_to_user(cls.user2, 1002)
+        # Ensure we operate on fresh model instances
+        self.skillfarm_audit.refresh_from_db()
 
-        cls.skill = create_skill_character(
-            character_id=cls.audit.character.character_id,
-            evetype_id=1,
-            skillpoints=500000,
-            trained_level=5,
-            active_level=5,
-        )
-        cls.skillsetup = create_skillsetup_character(
-            character_id=cls.audit.character.character_id, skillset=["skill1"]
-        )
-
-        cls.skill2 = create_skill_character(
-            character_id=cls.audit2.character.character_id,
-            evetype_id=2,
-            skillpoints=500000,
-            trained_level=5,
-            active_level=5,
-        )
-        cls.skillsetup2 = create_skillsetup_character(
-            character_id=cls.audit2.character.character_id, skillset=["skill2"]
-        )
-
-    def _set_notifiaction_status(self, audits, status):
-        for audit in audits:
-            audit.notification = status
-            audit.save()
-
-    def test_notifiaction_with_skillsetup_should_return_true(self, mock_audit_filter):
-        audits = [self.audit, self.audit2]
-        self._set_notifiaction_status(audits, True)
+        audits = [self.skillfarm_audit]
+        self._set_notification_status(audits, True)
 
         mock_audit_filter.return_value = audits
         # when
@@ -203,42 +183,23 @@ class TestCheckSkillfarmNotificationSuccess(TestCase):
             self.assertTrue(audit.notification_sent)
             self.assertIsNotNone(audit.last_notification)
 
-
-@patch(TASK_PATH + ".SkillFarmAudit.objects.filter", spec=True)
-@override_settings(
-    CELERY_ALWAYS_EAGER=True,
-    CELERY_EAGER_PROPAGATES_EXCEPTIONS=True,
-    APP_UTILS_OBJECT_CACHE_DISABLED=True,
-)
-class TestCheckSkillfarmNotificationError(TestCase):
-    @classmethod
-    def setUpClass(cls):
-        super().setUpClass()
-        load_allianceauth()
-        load_eveuniverse()
-
-        cls.user, cls.character_ownership = create_user_from_evecharacter_with_access(
-            1001
-        )
-        cls.audit = add_skillfarmaudit_character_to_user(cls.user, 1001)
-
-    def _set_notifiaction_status(self, audits, status):
-        for audit in audits:
-            audit.notification = status
-            audit.save()
-
     @patch(TASK_PATH + ".logger", spec=True)
     def test_notifiaction_no_main_should_return_false(
         self, mock_logger, mock_audit_filter
     ):
-        audits = [self.audit]
-        self._set_notifiaction_status(audits, True)
+        """
+        Test should not send notification if no main character is found.
+        """
+        audits = [self.skillfarm_audit]
+        self._set_notification_status(audits, True)
 
-        userprofile = UserProfile.objects.get(user=self.user)
-        userprofile.main_character = None
-        userprofile.save()
-        self.character_ownership.delete()
-        self.audit.refresh_from_db()
+        # Delete the main character and clear the relation
+        self.user.profile.main_character = None
+        self.user.profile.save()
+        self.user.profile.refresh_from_db()
+
+        # Ensure we operate on fresh model instances (no cached relations)
+        audits = [SkillFarmAudit.objects.get(pk=audit.pk) for audit in audits]
 
         mock_audit_filter.return_value = audits
         # when
@@ -249,7 +210,7 @@ class TestCheckSkillfarmNotificationError(TestCase):
             self.assertIsNone(audit.last_notification)
             mock_logger.warning.assert_called_once_with(
                 "Main Character not found for %s, skipping notification",
-                self.audit.character.character_name,
+                self.skillfarm_audit.character.character_name,
             )
 
 
@@ -257,22 +218,29 @@ class TestCheckSkillfarmNotificationError(TestCase):
 @override_settings(
     CELERY_ALWAYS_EAGER=True,
     CELERY_EAGER_PROPAGATES_EXCEPTIONS=True,
-    APP_UTILS_OBJECT_CACHE_DISABLED=True,
 )
-class TestSkillfarmPrices(TestCase):
+class TestSkillfarmPrices(SkillFarmTestCase):
     @classmethod
     def setUpClass(cls):
         super().setUpClass()
-        load_allianceauth()
-        load_eveuniverse()
 
-        cls.price = create_evetypeprice(3, buy=100, sell=200, updated_at=timezone.now())
-        cls.price2 = create_evetypeprice(
-            2, buy=300, sell=400, updated_at=timezone.now()
+        cls.price = create_eve_type_price(
+            name="TestPrice",
+            eve_type_id=44992,
+            buy=100,
+            sell=200,
+            updated_at=timezone.now(),
+        )
+        cls.price2 = create_eve_type_price(
+            name="TestPrice2",
+            eve_type_id=40519,
+            buy=300,
+            sell=400,
+            updated_at=timezone.now(),
         )
 
         cls.json = {
-            3: {
+            "44992": {
                 "buy": {"percentile": 100},
                 "sell": {"percentile": 200},
             }
@@ -283,6 +251,9 @@ class TestSkillfarmPrices(TestCase):
     def test_update_prices_should_update_nothing(
         self, mock_logger, mock_prices, mock_requests
     ):
+        """
+        Test should not update prices if no prices exist.
+        """
         mock_prices.return_value = []
         mock_response = mock_requests.return_value
         mock_response.json.return_value = self.json
@@ -293,16 +264,24 @@ class TestSkillfarmPrices(TestCase):
         mock_logger.info.assert_called_once_with("No Prices to update")
 
     def test_should_update_prices(self, mock_requests):
+        """
+        Test should update existing prices.
+        """
         mock_response = mock_requests.return_value
         mock_response.json.return_value = self.json
+        old_updated_at = self.price.updated_at
         # when
         tasks.update_all_prices()
+        self.price.refresh_from_db()
         # then
-        self.assertAlmostEqual(self.price.buy, 100)
-        self.assertAlmostEqual(self.price.sell, 200)
-        self.assertIsNotNone(self.price.updated_at)
+        self.assertEqual(self.price.buy, 100)
+        self.assertEqual(self.price.sell, 200)
+        self.assertGreater(self.price.updated_at, old_updated_at)
 
     def test_update_prices_should_only_update_existing(self, mock_requests):
+        """
+        Test should only update existing prices.
+        """
         mock_response = mock_requests.return_value
         changed_json = self.json.copy()
         changed_json.update(
@@ -316,16 +295,18 @@ class TestSkillfarmPrices(TestCase):
         mock_response.json.return_value = changed_json
         # when
         tasks.update_all_prices()
+        self.price.refresh_from_db()
         # then
-        self.assertAlmostEqual(self.price.buy, 100)
-        self.assertAlmostEqual(self.price.sell, 200)
-        self.assertIsNotNone(self.price.updated_at)
+        self.assertIsNone(EveTypePrice.objects.filter(eve_type__id=4).first())
 
     @patch(TASK_PATH + ".EveTypePrice.objects.bulk_update", spec=True)
     @patch(TASK_PATH + ".logger", spec=True)
     def test_update_prices_should_raise_exception(
         self, mock_logger, mock_bulk_update, mock_requests
     ):
+        """
+        Test should log error if bulk_update raises an exception.
+        """
         mock_response = mock_requests.return_value
         mock_response.json.return_value = self.json
         error_instance = Error("Error")