+++ skillfarm/tests/views/test_delete_character.py
@@ -5,35 +5,33 @@ import json
 from http import HTTPStatus
 
 # Django
-from django.test import RequestFactory, TestCase
 from django.urls import reverse
 
 # AA Skillfarm
-from skillfarm.tests.testdata.allianceauth import load_allianceauth
-from skillfarm.tests.testdata.eveuniverse import load_eveuniverse
-from skillfarm.tests.testdata.skillfarm import (
-    create_skillfarm_character,
-    create_user_from_evecharacter_with_access,
-)
+from skillfarm.tests import SkillFarmTestCase
+from skillfarm.tests.testdata.utils import create_skillfarm_character_from_user
 from skillfarm.views import delete_character
 
 MODULE_PATH = "skillfarm.views"
 
 
-class TestDeleteCharacterView(TestCase):
+class TestDeleteCharacterView(SkillFarmTestCase):
+    """
+    Test Delete Character Ajax Response.
+    """
+
     @classmethod
     def setUpClass(cls):
         super().setUpClass()
-        load_allianceauth()
-        load_eveuniverse()
 
-        cls.factory = RequestFactory()
-        cls.audit = create_skillfarm_character(1001)
-        cls.user = cls.audit.character.character_ownership.user
-        cls.no_audit_user, _ = create_user_from_evecharacter_with_access(1002)
+        cls.skillfarm_audit = create_skillfarm_character_from_user(cls.user)
+        cls.skillfarm_audit_2 = create_skillfarm_character_from_user(cls.superuser)
 
     def test_delete_character(self):
-        character_id = self.audit.character.character_id
+        """
+        Test should delete own Character successfully.
+        """
+        character_id = self.skillfarm_audit.character.character_id
         form_data = {
             "character_id": character_id,
             "confirm": "yes",
@@ -52,19 +50,21 @@ class TestDeleteCharacterView(TestCase):
         self.assertTrue(response_data["success"])
         self.assertEqual(response_data["message"], "Gneuten successfully deleted")
 
-    def test_delete_character_no_audit(self):
+    def test_delete_character_no_permission(self):
+        """
+        Test should prevent deleting Character that are not owned by the user.
+        """
         form_data = {
-            "character_id": 1001,
+            "character_id": 1003,
             "confirm": "yes",
         }
 
         request = self.factory.post(
-            reverse("skillfarm:delete_character", args=[1001]), data=form_data
+            reverse("skillfarm:delete_character", args=[1003]), data=form_data
         )
-        request.user = self.no_audit_user
+        request.user = self.user
 
-        response = delete_character(request, character_id=1001)
-
+        response = delete_character(request, character_id=1003)
         response_data = json.loads(response.content)
 
         self.assertEqual(response.status_code, HTTPStatus.FORBIDDEN)
@@ -72,12 +72,14 @@ class TestDeleteCharacterView(TestCase):
         self.assertEqual(response_data["message"], "Permission Denied")
 
     def test_delete_character_invalid(self):
+        """
+        Test should prevent deleting Character with invalid form data.
+        """
         request = self.factory.post(
-            reverse("skillfarm:delete_character", args=[1001]), data=None
+            reverse("skillfarm:delete_character", args=[1003]), data=None
         )
-        request.user = self.no_audit_user
-
-        response = delete_character(request, character_id=1001)
+        request.user = self.user
+        response = delete_character(request, character_id=1003)
 
         response_data = json.loads(response.content)