+++ skillfarm/tests/views/test_skill_calculator.py
@@ -2,47 +2,50 @@
 
 # Standard Library
 from http import HTTPStatus
-from unittest.mock import Mock, patch
 
 # Django
-from django.test import RequestFactory, TestCase
 from django.urls import reverse
 from django.utils import timezone
 
 # AA Skillfarm
-from skillfarm.tests.testdata.allianceauth import load_allianceauth
-from skillfarm.tests.testdata.eveuniverse import load_eveuniverse
-from skillfarm.tests.testdata.skillfarm import (
-    create_evetypeprice,
-    create_skillfarm_character,
-    create_user_from_evecharacter_with_access,
+from skillfarm.tests import SkillFarmTestCase
+from skillfarm.tests.testdata.utils import (
+    create_eve_type_price,
 )
 from skillfarm.views import skillfarm_calc
 
 MODULE_PATH = "skillfarm.views"
 
 
-class TestSkillFarmCalculator(TestCase):
+class TestSkillFarmCalculator(SkillFarmTestCase):
+    """Test Skill Calculator View."""
+
     @classmethod
     def setUpClass(cls):
         super().setUpClass()
-        load_allianceauth()
-        load_eveuniverse()
 
-        cls.factory = RequestFactory()
-        cls.user, _ = create_user_from_evecharacter_with_access(1001)
-
-        cls.plex = create_evetypeprice(
-            44992, buy=100, sell=200, updated_at=timezone.now()
+        cls.plex = create_eve_type_price(
+            name="PLEX", eve_type_id=44992, buy=100, sell=200, updated_at=timezone.now()
         )
-        cls.skillinjector = create_evetypeprice(
-            40520, buy=300, sell=400, updated_at=timezone.now()
+        cls.skillinjector = create_eve_type_price(
+            name="Skill Injector",
+            eve_type_id=40520,
+            buy=300,
+            sell=400,
+            updated_at=timezone.now(),
         )
-        cls.extractor = create_evetypeprice(
-            40519, buy=500, sell=600, updated_at=timezone.now()
+        cls.extractor = create_eve_type_price(
+            name="Skill Extractor",
+            eve_type_id=40519,
+            buy=500,
+            sell=600,
+            updated_at=timezone.now(),
         )
 
     def test_skillcalculator_should_view_calc(self):
+        """
+        Test should render skill calculator view.
+        """
         # given
         request = self.factory.get(reverse("skillfarm:calc"))
         request.user = self.user
@@ -55,6 +58,9 @@ class TestSkillFarmCalculator(TestCase):
         self.assertContains(response, "Skill Extractor")
 
     def test_skillcalculator_should_view_calc_with_character_id(self):
+        """
+        Test should render skill calculator view with character id.
+        """
         # given
         request = self.factory.get(reverse("skillfarm:calc"))
         request.user = self.user