+++ skillfarm/thirdparty/charlink_hook.py
@@ -6,12 +6,11 @@ from django.contrib.auth.models import Permission
 from django.db.models import Exists, OuterRef
 
 # Alliance Auth
+from allianceauth.authentication.models import User
 from allianceauth.eveonline.models import EveCharacter
 
-# Alliance Auth (External Libs)
-from app_utils.allianceauth import users_with_permission
-
 # AA Skillfarm
+from skillfarm import __app_name__
 from skillfarm.app_settings import SKILLFARM_APP_NAME
 from skillfarm.models.skillfarmaudit import SkillFarmAudit
 from skillfarm.tasks import update_character
@@ -34,11 +33,21 @@ def _is_character_added_charaudit(character: EveCharacter):
 
 
 def _users_with_perms_charaudit():
-    return users_with_permission(
-        Permission.objects.get(
-            content_type__app_label="skillfarm", codename="basic_access"
-        )
+    permission = Permission.objects.get(
+        content_type__app_label=__app_name__, codename="basic_access"
     )
+
+    users_qs = (
+        permission.user_set.all()
+        | User.objects.filter(
+            groups__in=list(permission.group_set.values_list("pk", flat=True))
+        )
+        | User.objects.select_related("profile").filter(
+            profile__state__in=list(permission.state_set.values_list("pk", flat=True))
+        )
+        | User.objects.filter(is_superuser=True)
+    )
+    return users_qs.distinct()
 
 
 app_import = AppImport(